Explicitly exported all symbols. For gcc the macros need to be updated with the visibility attribute. (Nick Porcino) Made a sweep for forward class declarations and replaced with a "#include ImfForward.h". (Piotr Stanczyk)

+++ OpenEXR/IlmImf/ImfCRgbaFile.cpp
@@ -293,8 +293,8 @@ void
 ImfHeaderScreenWindowCenter (const ImfHeader *hdr, float *x, float *y)
 {
     const V2i &swc = header(hdr)->screenWindowCenter();
-    *x = swc.x;
-    *y = swc.y;
+    *x = (float) swc.x;
+    *y = (float) swc.y;
 }
